refactor(phase-7): flatten components and consolidate UI primitives in ui/
