+++ my-app/src/app/components/Experience.tsx
@@ -1,8 +1,8 @@
 "use client";
 import { useState } from "react";
 import { SectionHeader } from "./FeaturedWork";
-import type { ExperienceEntry } from "../../i18n/portfolio.types";
-import { usePortfolioData } from "../../i18n/usePortfolioData";
+import type { ExperienceEntry } from "../i18n/portfolio.types";
+import { usePortfolioData } from "../i18n/usePortfolioData";
 
 export default function Experience({ lang }: { lang: string }) {
   const { data, ready } = usePortfolioData(lang);
