+++ my-app/src/app/components/StackSection.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { SectionHeader } from "./FeaturedWork";
-import type { StackCategories } from "../../i18n/portfolio.types";
-import { usePortfolioData } from "../../i18n/usePortfolioData";
+import type { StackCategories } from "../i18n/portfolio.types";
+import { usePortfolioData } from "../i18n/usePortfolioData";
 
 export default function StackSection({ lang }: { lang: string }) {
   const { data, ready } = usePortfolioData(lang);
